Backend real com Supabase: banco Postgres, login Microsoft, RLS por papel e site FastAPI
Substitui o dashboard HTML estatico por um site com backend de verdade:
Postgres gerenciado + Auth + RLS via Supabase, autenticacao Microsoft
(Supabase Auth com provedor Azure), controle de acesso por papel
(admin/revisor/leitor) e um fluxo de revisao humana para OCs sinalizadas
na Central de alertas (marcar/desfazer revisao), com filtros por status e
data de emissao. Pipeline e relatorio HTML estatico migrados de SQLite para
Postgres/psycopg. Inclui pool de conexoes para reduzir latencia e o
blueprint de deploy (render.yaml) para a Fase 6.

+++ sql/queries/itens_mais_recorrentes.sql
@@ -4,7 +4,7 @@ SELECT
     i.descricao,
     COUNT(DISTINCT i.ordem_compra_id) AS numero_ocs,
     SUM(i.quantidade) AS quantidade_total,
-    ROUND(AVG(i.valor_unitario), 2) AS valor_unitario_medio,
+    ROUND(AVG(i.valor_unitario)::numeric, 2) AS valor_unitario_medio,
     SUM(i.valor_total) AS valor_total_acumulado
 FROM itens_oc i
 GROUP BY i.codigo_produto, i.descricao
