+++ sql/queries/valor_total_por_cliente.sql
@@ -5,7 +5,7 @@ SELECT
     c.uf,
     COUNT(oc.id) AS total_ocs,
     SUM(oc.valor_total) AS valor_total_acumulado,
-    ROUND(AVG(oc.valor_total), 2) AS valor_medio_por_oc
+    ROUND(AVG(oc.valor_total)::numeric, 2) AS valor_medio_por_oc
 FROM ordens_compra oc
 JOIN clientes c ON c.id = oc.cliente_id
 GROUP BY c.id
